backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
gunicorn
psycopg2-binary
PyJWT